requirements.txt
matplotlib==3.3.2
numpy==1.22.4
pandas==1.1.3
Pillow==9.2.0
streamlit==1.10.0
tensorflow==2.9.1
